requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asttokens==3.0.0
asyncer==0.0.7
attrs==25.1.0
backoff==2.2.1
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chainlit==2.5.5
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
griffe==1.7.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
Lazify==0.4.0
litellm==1.67.2
literalai==0.1.201
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
monotonic==1.6
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.29.0
nest-asyncio==1.6.0
numpy==2.2.3
openai==1.76.0
openai-agents==0.0.13
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.39.2
opentelemetry-instrumentation-anthropic==0.39.2
opentelemetry-instrumentation-bedrock==0.39.2
opentelemetry-instrumentation-chromadb==0.39.2
opentelemetry-instrumentation-cohere==0.39.2
opentelemetry-instrumentation-crewai==0.39.2
opentelemetry-instrumentation-google-generativeai==0.39.2
opentelemetry-instrumentation-groq==0.39.2
opentelemetry-instrumentation-haystack==0.39.2
opentelemetry-instrumentation-lancedb==0.39.2
opentelemetry-instrumentation-langchain==0.39.2
opentelemetry-instrumentation-llamaindex==0.39.2
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.39.2
opentelemetry-instrumentation-milvus==0.39.2
opentelemetry-instrumentation-mistralai==0.39.2
opentelemetry-instrumentation-ollama==0.39.2
opentelemetry-instrumentation-openai==0.39.2
opentelemetry-instrumentation-pinecone==0.39.2
opentelemetry-instrumentation-qdrant==0.39.2
opentelemetry-instrumentation-replicate==0.39.2
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.39.2
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.39.2
opentelemetry-instrumentation-transformers==0.39.2
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.39.2
opentelemetry-instrumentation-watsonx==0.39.2
opentelemetry-instrumentation-weaviate==0.39.2
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.52b1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
posthog==3.25.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.41.3
streamlit==1.42.2
syncer==2.0.3
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traceloop-sdk==0.39.2
traitlets==5.14.3
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uptrace==1.31.0
urllib3==2.3.0
uv==0.6.14
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==0.20.0
wcwidth==0.2.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
